Fixed bug stopping music from being sorted
For some reason the CreateDirectory function was throwing
an error related to the file not existing. So I cleaned up the code
a bit revolving around this which seemed to fix the issue using
Path.Combine.

Also added some console logging because apparently I didn't know
what that was before.